add base routes
