+++ app/javascript/components/Admin/Cylinder/Detail/index.js
@@ -1,0 +1,5 @@
+function CylinderDetail() {
+  return <p>index</p>;
+}
+
+export default CylinderDetail;
